indexer/pyproject.toml
[project]
name = "rpg-catalog-indexer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "PyMuPDF>=1.24",
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "httpx>=0.27",
    "anthropic>=0.40",
    "google-genai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "respx>=0.21",
    "reportlab>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
